requirements.txt
requests
beautifulsoup4
Pillow
fonttools
